feat: project-a nextjs 재설치
