+++ packages/project-a/src/app/favicon.ico
